日志保留时间修改为系统配置

+++ src/main/java/com/inmaytide/orbit/library/timer/AutoCleaningOperationLogJob.java
@@ -1,6 +1,7 @@
 package com.inmaytide.orbit.library.timer;
 
 import com.baomidou.mybatisplus.core.conditions.query.LambdaQueryWrapper;
+import com.inmaytide.orbit.commons.consts.Constants;
 import com.inmaytide.orbit.commons.domain.Tenant;
 import com.inmaytide.orbit.commons.log.domain.OperationLog;
 import com.inmaytide.orbit.commons.metrics.AbstractJob;
@@ -8,6 +9,7 @@ import com.inmaytide.orbit.commons.service.uaa.TenantService;
 import com.inmaytide.orbit.library.domain.OperationLogCleaning;
 import com.inmaytide.orbit.library.mapper.OperationLogCleaningMapper;
 import com.inmaytide.orbit.library.mapper.OperationLogMapper;
+import com.inmaytide.orbit.library.service.SystemPropertyService;
 import org.quartz.JobExecutionContext;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -36,6 +38,9 @@ public class AutoCleaningOperationLogJob extends AbstractJob {
     @Autowired
     private OperationLogMapper operationLogMapper;
 
+    @Autowired
+    private SystemPropertyService propertyService;
+
     @Override
     public Logger getLogger() {
         return LOG;
@@ -47,17 +52,16 @@ public class AutoCleaningOperationLogJob extends AbstractJob {
     }
 
     @Override
-    public boolean fireImmediatelyWhenServiceStartup() {
-        return true;
-    }
-
-    @Override
     protected void exec(JobExecutionContext context) {
-        List<Tenant> tenants = tenantService.all().stream().filter(e -> e.getOperationLogRetentionTime() != -1).toList();
+        List<Tenant> tenants = tenantService.all();
         for (Tenant e : tenants) {
+            Integer retentionTime = propertyService.getIntValue(e.getId(), Constants.SPK_OP_LOG_RETENTION_TIME_IN_DAYS);
+            if (retentionTime == null || retentionTime == -1) {
+                continue;
+            }
             LambdaQueryWrapper<OperationLog> wrapper = new LambdaQueryWrapper<>();
             wrapper.eq(OperationLog::getTenantId, e.getId());
-            wrapper.apply("operation_time < TIMESTAMPADD(DAY, {0}, now())", e.getOperationLogRetentionTime());
+            wrapper.apply("operation_time < TIMESTAMPADD(DAY, {0}, now())", retentionTime);
             int affected = operationLogMapper.delete(wrapper);
             if (affected > 0) {
                 OperationLogCleaning entity = new OperationLogCleaning();
